+++ templates/userPanel.html
@@ -40,7 +40,7 @@
 
     </div>
     <div class="d-flex row flex-column pt-2">
-        <form action="{{ url_for('operation_controller.add_file')}}" method="POST" enctype="multipart/form-data">
+        <form action="{{ url_for('user_controller.add_file')}}" method="POST" enctype="multipart/form-data">
             <div class="table-responsive">
                 <table id="files-table" class="table table-hover table-dark mb-0">
                     <thead>
